+++ lib/node/runner/api/resource/operation.go
