+++ lib/node/runner/node_runner.go
@@ -19,7 +19,7 @@ import (
 	"boscoin.io/sebak/lib/consensus/round"
 	"boscoin.io/sebak/lib/error"
 	"boscoin.io/sebak/lib/network"
-	"boscoin.io/sebak/lib/network/api"
+	"boscoin.io/sebak/lib/node/runner/api"
 	"boscoin.io/sebak/lib/node"
 	"boscoin.io/sebak/lib/storage"
 	"boscoin.io/sebak/lib/transaction"
